feat: Support log rotate for rclone sync
This commit adds support for log rotate to the rclone sync script, which will help keep the log files from getting too large.

The maximum number of log files to keep is configurable through the `$maximumLogFiles` variable, and the default is 15.

The log files will be rotated based on the last write time, with the oldest log files being deleted first.