refactor(rclone): rename "constants and variables" to "variables"
This commit renames the "constants and variables" section in the rclone-sync.ps1 script to "variables". The "constants and variables" section contains variables used by the script, but it does not contain any constants. Therefore, the new name "variables" is more accurate.

This change does not affect the functionality of the script.