feat(cli): support passing path to rclone executable
This commit adds the ability to pass the path to the rclone executable to the `rclone-sync` script.
This is useful for users who have installed rclone in a non-standard location.

Additionally, the default value of the `RclonePath` parameter has been set to `rclone`.

BREAKING CHANGE: The `RclonePath` parameter is now required.